Move data related files to data folder in client
