+++ client/src/data/indoorPoisLogic.js
